requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
python-dotenv
pydantic
anthropic
ollama
pillow
pytesseract
